rename dev server
